mod: 移动check_torrent_duplicate函数位置
