+++ backend/src/module/searcher/searcher.py
@@ -1,13 +1,11 @@
 import json
 from typing import TypeAlias
 
-from module.conf import settings
 from module.models import Bangumi, RSSItem, Torrent
 from module.network import RequestContent
+from module.manager.torrent import TorrentManager
 from module.rss import RSSAnalyser
 
-from module.manager.renamer import Renamer
-from module.parser import TitleParser
 from .provider import search_url
 
 SEARCH_KEY = [
@@ -22,7 +20,7 @@ SEARCH_KEY = [
 BangumiJSON: TypeAlias = str
 
 
-class SearchTorrent(RequestContent, RSSAnalyser, TitleParser, Renamer):
+class SearchTorrent(RequestContent, RSSAnalyser, TorrentManager):
     def search_torrents(self, rss_item: RSSItem) -> list[Torrent]:
         return self.get_torrents(rss_item.url)
         # torrents = self.get_torrents(rss_item.url)
@@ -51,25 +49,6 @@ class SearchTorrent(RequestContent, RSSAnalyser, TitleParser, Renamer):
         keywords = [getattr(data, key) for key in SEARCH_KEY if getattr(data, key)]
         url = search_url(site, keywords)
         return url
-
-    def check_torrent_duplicate(self, bangumi: Bangumi, torrent: Torrent):
-        rename_method = settings.bangumi_manage.rename_method
-        # 查找视频是否包含关键字
-        keywords = ['内嵌','外挂','封装','MKV','内挂']
-        suffix = 'mp4'
-        for item1 in keywords:
-            if(item1 in torrent.name):
-                suffix = 'mkv'
-        # 解析视频
-        ep = self.torrent_parser(
-            torrent_path=f"{torrent.name}.{suffix}"
-        )
-        if(ep):
-            ep_name = self.gen_path(ep, bangumi.official_title, method=rename_method)
-            torrent.bangumi_id = bangumi.id
-            torrent.save_path = f'{bangumi.save_path}/{ep_name}'
-            return torrent
-        return False
     
     def search_season(self, data: Bangumi, site: str = "mikan") -> list[Torrent]:
         rss_item = self.special_url(data, site)
